requirements.txt
streamlit
numpy
pandas
scipy
Pillow
openpyxl
xlsxwriter
matplotlib
japanize-matplotlib
seaborn
scikit-learn
plotly
factor_analyzer
ipython
tqdm
seaborn
vistats
wordcloud
networkx
mecab-python3
unidic-lite
nlplot[all]
Counter
cufflinks
statsmodels
pingouin
janome
factor_analyzer